refactor(database): refactor of CLI and collection operations

+++ scanbackup/infrastructure/persistence/mongodb/connections/database.py
@@ -1,10 +1,16 @@
 from typing import Any, Dict
 from pathlib import Path
 from pymongo import MongoClient
-from pymongo.collection import Collection
 from scanbackup.shared import (
+    ModuleSystem,
+    SchemaDBAvailable,
+    Configuration,
+    DatabaseConfigModel,
+    LayerConfigModel,
+    LayerNotDefined,
     MongoDatabaseError,
     MongoConnectionError,
+    MongoCollectionNotFoundError,
     MongoCreateCollectionError,
     MongoDeleteCollectionError,
     MongoImportCollectionError,
@@ -15,6 +21,7 @@ from scanbackup.shared import (
 )
 from scanbackup.infrastructure.persistence.mongodb.constants.collection import (
     MongoCollectionName,
+    SuffixCollectionName,
 )
 from scanbackup.infrastructure.persistence.mongodb.collections.records.bbip.traffic import (
     BBIPCollection,
@@ -37,14 +44,16 @@ from scanbackup.infrastructure.persistence.mongodb.collections.summaries.bbip.da
 
 
 class MongoDatabase:
+    _instance: "MongoDatabase | None" = None
+    _config: DatabaseConfigModel
     _client: MongoClient[Any]
-    _name_db: str
     _uri: str
     connected: bool = False
 
-    def __init__(self, uri: str) -> None:
-        self._uri = uri
-        self.open_connection(uri)
+    def __new__(cls) -> "MongoDatabase":
+        if not cls._instance:
+            cls._instance = super(MongoDatabase, cls).__new__(cls)
+        return cls._instance
 
     def _check_collection(self, name: str) -> bool:
         """Check if the collection exists."""
@@ -52,11 +61,15 @@ class MongoDatabase:
         collection_list = db.list_collection_names()
         return name in collection_list
 
+    def set_uri(self, config: DatabaseConfigModel) -> None:
+        self._config = config
+        self._uri = f"mongodb://{self._config.user}:{self._config.password}@{self._config.host}:{self._config.port}"
+
     def get_uri(self) -> str:
         """Get the URI database."""
         return self._uri
 
-    def open_connection(self, uri: str | None = None) -> None:
+    def open_connection(self) -> None:
         """Open a connection to the database.
 
         :param uri: URI database
@@ -64,11 +77,7 @@ class MongoDatabase:
         """
         try:
             if not self.connected:
-                if not uri:
-                    uri = self._uri
-                name_db = uri.split("/")[-1]
-                self._name_db = name_db
-                self._client = MongoClient(uri)
+                self._client = MongoClient(self._uri)
         except Exception as error:
             self.connected = False
             raise MongoConnectionError(error=error, extra_msg="Fallo al abrir conexión")
@@ -82,18 +91,6 @@ class MongoDatabase:
         """
         return self._client
 
-    def get_cursor(self, table: str | None = None) -> Collection[Any]:
-        """Get a cursor to the database.
-
-        :param table: Collection name
-        :type table: str | None
-        :return MongoClient[Any]: Connection in the database collection
-        """
-        if table:
-            return self._client[self._name_db][table]
-        else:
-            return self._client[self._name_db].get_default_collection()
-
     def close_connection(self) -> None:
         """Close the connection to the database."""
         try:
@@ -104,52 +101,38 @@ class MongoDatabase:
                 error=error, extra_msg="Fallo al cerrar conexión"
             )
 
-    def initialize(self) -> None:
+    def create_collections(self, config: LayerConfigModel) -> None:
         """Create all collections and schemas in the database."""
         try:
             self.open_connection(self._uri)
-            db = self._client[self._name_db]
+            db = self._client[self._config.name]
+
             if not self._check_collection(MongoCollectionName.BBIP_SOURCES):
                 BBIPSourceCollection.create(database=db)
-            if not self._check_collection(MongoCollectionName.BORDE):
-                BBIPCollection.create(
-                    name_collection=MongoCollectionName.BORDE, database=db
-                )
-            if not self._check_collection(MongoCollectionName.BRAS):
-                BBIPCollection.create(
-                    name_collection=MongoCollectionName.BRAS, database=db
-                )
-            if not self._check_collection(MongoCollectionName.CACHING):
-                BBIPCollection.create(
-                    name_collection=MongoCollectionName.CACHING, database=db
-                )
-            if not self._check_collection(MongoCollectionName.RAI):
-                BBIPCollection.create(
-                    name_collection=MongoCollectionName.RAI, database=db
-                )
-            if not self._check_collection(MongoCollectionName.IXP):
-                BBIPCollection.create(
-                    name_collection=MongoCollectionName.IXP, database=db
-                )
-            if not self._check_collection(MongoCollectionName.DINT):
-                BBIPCollection.create(
-                    name_collection=MongoCollectionName.DINT, database=db
-                )
-            if not self._check_collection(MongoCollectionName.DIST):
-                BBIPCollection.create(
-                    name_collection=MongoCollectionName.DIST, database=db
-                )
+
+            layers_bbip = config.bbip.names
+            for layer in layers_bbip:
+                layer = layer.upper()
+                name_collection = layer + "_" + SuffixCollectionName.BBIP_HISTORIES
+                if not self._check_collection(name_collection):
+                    BBIPCollection.create(name_collection=name_collection, database=db)
+
             if not self._check_collection(MongoCollectionName.BBIP_DAILY_SUMMARY):
                 BBIPDailySummaryCollection.create(database=db)
-            if not self._check_collection(MongoCollectionName.IP_SOURCES):
-                IPSourceCollection.create(database=db)
 
-            if not self._check_collection(MongoCollectionName.IP_BRAS):
-                IPCollection.create(
-                    name_collection=MongoCollectionName.IP_BRAS, database=db
-                )
+            if not self._check_collection(name_collection):
+                IPSourceCollection.create(name_collection=name_collection, database=db)
+
+            layers_ip = config.ip.names
+            for layer in layers_ip:
+                layer = layer.upper()
+                name_collection = layer + "_" + SuffixCollectionName.IP_HISTORIES
+                if not self._check_collection(MongoCollectionName.IP_BRAS):
+                    IPCollection.create(name_collection=name_collection, database=db)
+
             if not self._check_collection(MongoCollectionName.IP_DAILY_SUMMARY):
                 IPDailySummaryCollection.create(database=db)
+
             self.close_connection()
         except MongoCreateCollectionError:
             raise
@@ -162,7 +145,8 @@ class MongoDatabase:
 
     def import_data(
         self,
-        name_collection: MongoCollectionName,
+        name_collection: str,
+        config: LayerConfigModel,
         filepath: str,
         delimiter: str | None = None,
     ) -> None:
@@ -170,16 +154,21 @@ class MongoDatabase:
         try:
             if not filepath.exists():
                 raise DatabaseDataNotFoundError()
+
+            if not self._check_collection(name_collection):
+                raise MongoCollectionNotFoundError(name_collection)
+
             if filepath.suffix != ".csv" or filepath.suffix != ".txt":
-                raise FileExtensionError(filepath=filepath, module="Mongo Database")
-            if name_collection is BBIPCollection.LAYERS_VALID:
+                raise FileExtensionError(filepath=filepath, module=ModuleSystem.MONGO)
+
+            if name_collection in config.bbip.names:
                 BBIPCollection.import_data(
                     name_collection=name_collection,
                     database=self._client,
                     input_path=filepath,
                     delimiter=delimiter,
                 )
-            elif name_collection is IPCollection.LAYERS_VALID:
+            elif name_collection is config.ip.names:
                 IPCollection.import_data(
                     name_collection=name_collection,
                     database=self._client,
@@ -211,8 +200,12 @@ class MongoDatabase:
                     delimiter=delimiter,
                 )
             else:
-                raise ValueError("Nombre de colección inválido")
+                raise LayerNotDefined(layer_name=name_collection)
+        except LayerNotDefined:
+            raise
         except FileExtensionError:
+            raise
+        except MongoCollectionNotFoundError:
             raise
         except MongoImportCollectionError:
             raise
@@ -227,8 +220,9 @@ class MongoDatabase:
 
     def export_data(
         self,
+        config: LayerConfigModel,
+        name_collection: str,
         dirpath: str | None = None,
-        name_collection: MongoCollectionName | None = None,
         delimiter: str | None = None,
         include_id: bool = True,
     ) -> None:
@@ -237,74 +231,27 @@ class MongoDatabase:
         else:
             dirpath = Path(dirpath)
         try:
-            if not name_collection or name_collection == MongoCollectionName.BORDE:
+            if name_collection and self._check_collection(name_collection):
+                raise MongoCollectionNotFoundError(name_collection)
+
+            if name_collection is config.bbip.names:
                 BBIPCollection.export_data(
-                    name_collection=MongoCollectionName.BORDE,
+                    name_collection=name_collection,
                     database=self._client,
-                    output_path=Path(dirpath / f"{MongoCollectionName.BORDE}.csv"),
+                    output_path=Path(dirpath / f"{name_collection}.csv"),
                     delimiter=delimiter,
                     include_id=include_id,
                 )
-            if not name_collection or name_collection == MongoCollectionName.BRAS:
-                BBIPCollection.export_data(
-                    name_collection=MongoCollectionName.BRAS,
-                    database=self._client,
-                    output_path=Path(dirpath / f"{MongoCollectionName.BRAS}.csv"),
-                    delimiter=delimiter,
-                    include_id=include_id,
-                )
-            if not name_collection or name_collection == MongoCollectionName.CACHING:
-                BBIPCollection.export_data(
-                    name_collection=MongoCollectionName.CACHING,
-                    database=self._client,
-                    output_path=Path(dirpath / f"{MongoCollectionName.CACHING}.csv"),
-                    delimiter=delimiter,
-                    include_id=include_id,
-                )
-            if not name_collection or name_collection == MongoCollectionName.RAI:
-                BBIPCollection.export_data(
-                    name_collection=MongoCollectionName.RAI,
-                    database=self._client,
-                    output_path=Path(dirpath / f"{MongoCollectionName.RAI}.csv"),
-                    delimiter=delimiter,
-                    include_id=include_id,
-                )
-            if not name_collection or name_collection == MongoCollectionName.DINT:
-                BBIPCollection.export_data(
-                    name_collection=MongoCollectionName.DINT,
-                    database=self._client,
-                    output_path=Path(dirpath / f"{MongoCollectionName.DINT}.csv"),
-                    delimiter=delimiter,
-                    include_id=include_id,
-                )
-            if not name_collection or name_collection == MongoCollectionName.DIST:
-                BBIPCollection.export_data(
-                    name_collection=MongoCollectionName.DIST,
-                    database=self._client,
-                    output_path=Path(dirpath / f"{MongoCollectionName.DIST}.csv"),
-                    delimiter=delimiter,
-                    include_id=include_id,
-                )
-            if not name_collection or name_collection == MongoCollectionName.IXP:
-                BBIPCollection.export_data(
-                    name_collection=MongoCollectionName.IXP,
-                    database=self._client,
-                    output_path=Path(dirpath / f"{MongoCollectionName.IXP}.csv"),
-                    delimiter=delimiter,
-                    include_id=include_id,
-                )
-            if not name_collection or name_collection == MongoCollectionName.IP_BRAS:
+
+            elif name_collection is config.ip.names:
                 IPCollection.export_data(
-                    name_collection=MongoCollectionName.IP_BRAS,
+                    name_collection=name_collection,
                     database=self._client,
-                    output_path=Path(dirpath / f"{MongoCollectionName.IP_BRAS}.csv"),
+                    output_path=Path(dirpath / f"{name_collection}.csv"),
                     delimiter=delimiter,
                     include_id=include_id,
                 )
-            if (
-                not name_collection
-                or name_collection == MongoCollectionName.BBIP_SOURCES
-            ):
+            elif name_collection == MongoCollectionName.BBIP_SOURCES:
                 BBIPSourceCollection.export_data(
                     database=self._client,
                     output_path=Path(
@@ -313,17 +260,14 @@ class MongoDatabase:
                     delimiter=delimiter,
                     include_id=include_id,
                 )
-            if not name_collection or name_collection == MongoCollectionName.IP_SOURCES:
+            elif name_collection == MongoCollectionName.IP_SOURCES:
                 IPSourceCollection.export_data(
                     database=self._client,
                     output_path=Path(dirpath / f"{MongoCollectionName.IP_SOURCES}.csv"),
                     delimiter=delimiter,
                     include_id=include_id,
                 )
-            if (
-                not name_collection
-                or name_collection == MongoCollectionName.BBIP_DAILY_SUMMARY
-            ):
+            elif name_collection == MongoCollectionName.BBIP_DAILY_SUMMARY:
                 BBIPDailySummaryCollection.export_data(
                     database=self._client,
                     output_path=Path(
@@ -332,10 +276,7 @@ class MongoDatabase:
                     delimiter=delimiter,
                     include_id=include_id,
                 )
-            if (
-                not name_collection
-                or name_collection == MongoCollectionName.IP_DAILY_SUMMARY
-            ):
+            elif name_collection == MongoCollectionName.IP_DAILY_SUMMARY:
                 IPDailySummaryCollection.export_data(
                     database=self._client,
                     output_path=Path(
@@ -345,43 +286,60 @@ class MongoDatabase:
                     include_id=include_id,
                 )
             else:
-                raise ValueError("Nombre de colección inválido")
+                raise LayerNotDefined(layer_name=name_collection)
+        except LayerNotDefined:
+            raise
+        except MongoCollectionNotFoundError:
+            raise
         except MongoExportCollectionError:
             raise
-        except ValueError:
-            if name_collection:
-                raise MongoExportCollectionError(
-                    name=name_collection, extra_msg="Nombre de la colección desconocido"
-                )
-            else:
-                raise
         except Exception as error:
             raise MongoDatabaseError(
                 error=error, message="Error inesperado al importar la data"
             )
 
-    def drop(self, force: bool = False) -> None:
+    def drop(self, config: LayerConfigModel, force: bool = False) -> None:
         """Deletes all collections in the database."""
         try:
             self.open_connection(self._uri)
-            db = self._client[self._name_db]
+            db = self._client[self._config.name]
 
-            if not force:
-                self.export_data()
+            for layer in config.bbip.names:
+                layer = layer.upper() + "_" + SuffixCollectionName.BBIP_HISTORIES
+                if not self._check_collection(layer):
+                    continue
+                if not force:
+                    self.export_data(config, layer)
+                BBIPCollection.delete(name_collection=layer, database=db)
 
-            BBIPCollection.delete(MongoCollectionName.BORDE, db)
-            BBIPCollection.delete(MongoCollectionName.BRAS, db)
-            BBIPCollection.delete(MongoCollectionName.CACHING, db)
-            BBIPCollection.delete(MongoCollectionName.RAI, db)
-            BBIPCollection.delete(MongoCollectionName.DINT, db)
-            BBIPCollection.delete(MongoCollectionName.DIST, db)
-            BBIPCollection.delete(MongoCollectionName.IXP, db)
-            BBIPDailySummaryCollection.delete(db)
-            BBIPSourceCollection.delete(db)
+            if self._check_collection(MongoCollectionName.BBIP_DAILY_SUMMARY):
+                if not force:
+                    self.export_data(config, MongoCollectionName.BBIP_DAILY_SUMMARY)
+                BBIPDailySummaryCollection.delete(db)
 
-            IPCollection.delete(MongoCollectionName.IP_BRAS, db)
-            IPDailySummaryCollection.delete(db)
-            IPSourceCollection.delete(db)
+            if self._check_collection(MongoCollectionName.BBIP_SOURCES):
+                if not force:
+                    self.export_data(config, MongoCollectionName.BBIP_SOURCES)
+                BBIPSourceCollection.delete(db)
+
+            for layer in config.bbip.names:
+                layer = layer.upper() + "_" + SuffixCollectionName.IP_HISTORIES
+                if not self._check_collection(layer):
+                    continue
+                if not force:
+                    self.export_data(config, layer)
+                IPCollection.delete(name_collection=layer, database=db)
+
+            if self._check_collection(MongoCollectionName.IP_DAILY_SUMMARY):
+                if not force:
+                    self.export_data(config, MongoCollectionName.IP_DAILY_SUMMARY)
+                IPDailySummaryCollection.delete(db)
+
+            if self._check_collection(MongoCollectionName.IP_SOURCES):
+                if not force:
+                    self.export_data(config, MongoCollectionName.IP_SOURCES)
+                IPSourceCollection.delete(db)
+
             self.close_connection()
         except MongoExportCollectionError:
             raise
